remove the rest, wait helper function in eventtest to prevent use in graphapi
